fix: use float64 for SHAP predict wrapper to match StructureNetwork weights
StructureNetwork stores weights as float64 tensors. The model_predict
wrapper was casting input to float32, causing dtype mismatch on matmul.

+++ explaneat/api/routes/evidence.py
@@ -563,7 +563,7 @@ async def compute_shap(
 
             def model_predict(x):
                 import torch
-                x = torch.as_tensor(x, dtype=torch.float32)
+                x = torch.as_tensor(x, dtype=torch.float64)
                 out = struct_net.forward(x).detach().numpy()
                 # Flatten single-output to 1D for SHAP
                 if out.ndim == 2 and out.shape[1] == 1:
